+++ CHANGELOG.md
@@ -8,6 +8,7 @@ New features:
 - Linux: Subscribe to capture device control events for volume, sample rate and format changes.
 - Linux: Optionally select Alsa sample format automatically.
 - Improved controller for rate adjustment.
+- Command line options fo setting aux volume and mute.
 Changes:
 - Rename `File` capture device to `RawFile`.
 - Filter pipeline steps take a list of channels to filter instead of a single one.